requirements.txt
# Data handling
pandas
duckdb
pandasql

# Google Gemini API
google-genai

